refactor: extract block tree and component instance utilities
