+++ frontend/src/block.ts
@@ -1,11 +1,16 @@
 import useCanvasStore from "@/stores/canvasStore";
 import useComponentStore from "@/stores/componentStore";
+import { findBlockInTree, resetBlock } from "@/utils/block/tree";
+import {
+	extendWithComponent,
+	rebuildWithComponent,
+	resetWithComponent,
+	syncBlockWithComponent,
+} from "@/utils/block/componentInstance";
 import {
 	addPxToNumber,
 	cssUrl,
 	dataURLtoFile,
-	deepEqual,
-	diffArray,
 	generateId,
 	getBlockCopy,
 	getBlockInstance,
@@ -20,7 +25,7 @@ import {
 } from "@/utils/helpers";
 import { Editor } from "@tiptap/vue-3";
 import { clamp } from "@vueuse/core";
-import { computed, nextTick, reactive, toRaw } from "vue";
+import { computed, nextTick, reactive } from "vue";
 
 const TEXT_ELEMENTS = new Set([
 	"span",
@@ -864,20 +869,8 @@ class Block implements BlockOptions {
 			syncBlockWithComponent(this, this, this.extendedFromComponent as string, component.children);
 		}
 	}
-	rebuildWithComponent(
-		componentId: string,
-		newComponentChildren: Block[],
-		oldComponentChildren: Block[],
-	) {
-		const oldChildrenByRefId = indexChildrenByRefId(this.children);
-		const oldComponentChildrenByBlockId = indexChildrenByBlockId(oldComponentChildren);
-		rebuildWithComponent(
-			this,
-			componentId,
-			newComponentChildren,
-			oldChildrenByRefId,
-			oldComponentChildrenByBlockId,
-		);
+	rebuildWithComponent(componentId: string, newComponentChildren: Block[], oldComponentChildren: Block[]) {
+		rebuildWithComponent(this, componentId, newComponentChildren, oldComponentChildren);
 	}
 	resetChanges(resetChildren: boolean = false) {
 		resetBlock(this, resetChildren);
@@ -1081,255 +1074,6 @@ class Block implements BlockOptions {
 		if (!componentRoot) return;
 		componentRoot.props = props;
 	}
-}
-
-function extendWithComponent(
-	block: Block | BlockOptions,
-	extendedFromComponent: string | undefined,
-	componentChildren: Block[],
-	resetOverrides: boolean = true,
-) {
-	resetBlock(block, false, resetOverrides);
-	block.children?.forEach((child, index) => {
-		child.isChildOfComponent = extendedFromComponent;
-		let componentChild = componentChildren[index];
-		if (child.extendedFromComponent) {
-			const component = child.referenceComponent;
-			child.referenceBlockId = componentChild.blockId;
-			extendWithComponent(child, child.extendedFromComponent, component.children, false);
-		} else if (componentChild) {
-			child.referenceBlockId = componentChild.blockId;
-			extendWithComponent(child, extendedFromComponent, componentChild.children, resetOverrides);
-		}
-	});
-}
-
-function resetWithComponent(
-	block: Block | BlockOptions,
-	extendedWithComponent: string,
-	componentChildren: Block[],
-	resetOverrides: boolean = true,
-) {
-	block = toRaw(block);
-	resetBlock(block, true, resetOverrides);
-	block.children?.splice(0, block.children.length);
-	componentChildren.forEach((componentChild) => {
-		const blockComponent = getBlockCopy(componentChild);
-		blockComponent.isChildOfComponent = extendedWithComponent;
-		blockComponent.referenceBlockId = componentChild.blockId;
-		if (!blockComponent.extendedFromComponent) {
-			blockComponent.componentVersion = block.componentVersion;
-		}
-		const childBlock = block.addChild(blockComponent, null, false);
-		if (componentChild.extendedFromComponent) {
-			const component = childBlock.referenceComponent;
-			resetWithComponent(childBlock, componentChild.extendedFromComponent, component.children, false);
-		} else {
-			resetWithComponent(childBlock, extendedWithComponent, componentChild.children, resetOverrides);
-		}
-	});
-}
-
-function syncBlockWithComponent(
-	parentBlock: Block,
-	block: Block,
-	componentName: string,
-	componentChildren: Block[],
-) {
-	componentChildren.forEach((componentChild, index) => {
-		const blockExists = findComponentBlock(componentChild.blockId, parentBlock.children);
-		if (!blockExists) {
-			const blockComponent = getBlockCopy(componentChild);
-			blockComponent.isChildOfComponent = componentName;
-			blockComponent.referenceBlockId = componentChild.blockId;
-			resetBlock(blockComponent);
-			resetWithComponent(blockComponent, componentName, componentChild.children);
-			block.addChild(blockComponent, index, false);
-		}
-	});
-
-	block.children.forEach((child) => {
-		const componentChild = componentChildren.find((c) => c.blockId === child.referenceBlockId);
-		if (componentChild) {
-			syncBlockWithComponent(parentBlock, child, componentName, componentChild.children);
-		}
-	});
-}
-
-const mergeOverrideMap = (
-	src: Record<string, any> | undefined,
-	oldComp: Record<string, any> | undefined,
-	dst: Record<string, any>,
-) => {
-	for (const key of Object.keys(src || {})) {
-		if (!deepEqual(src?.[key], oldComp?.[key])) {
-			dst[key] = src?.[key];
-		}
-	}
-};
-
-// Props have a schema defined by the component: only keys present in the new version's props
-// are valid. User overrides survive only for keys the new version still declares; props the
-// new version dropped are removed even if the user overrode them.
-const mergeOverrideProps = (
-	src: Record<string, any> | undefined,
-	oldComp: Record<string, any> | undefined,
-	dstProps: Record<string, any>,
-): Record<string, any> => {
-	const result: Record<string, any> = {};
-	for (const key of Object.keys(dstProps)) {
-		if (src && key in src && !deepEqual(src[key], oldComp?.[key])) {
-			result[key] = src[key];
-		} else {
-			result[key] = dstProps[key];
-		}
-	}
-	return result;
-};
-
-const copyUserOverrides = (src: Block, dst: Block, oldComponentChild: Block) => {
-	mergeOverrideMap(src.baseStyles, oldComponentChild.baseStyles, dst.baseStyles);
-	mergeOverrideMap(src.rawStyles, oldComponentChild.rawStyles, dst.rawStyles);
-	mergeOverrideMap(src.mobileStyles, oldComponentChild.mobileStyles, dst.mobileStyles);
-	mergeOverrideMap(src.tabletStyles, oldComponentChild.tabletStyles, dst.tabletStyles);
-	mergeOverrideMap(src.attributes, oldComponentChild.attributes, dst.attributes);
-	mergeOverrideMap(src.customAttributes, oldComponentChild.customAttributes, dst.customAttributes);
-	dst.props = mergeOverrideProps(src.props, oldComponentChild.props, dst.props || {});
-	if (src.classes) {
-		dst.classes = [...(dst.classes || []), ...diffArray(src.classes, oldComponentChild.classes)];
-	}
-	if (src.dynamicValues) {
-		dst.dynamicValues = [
-			...(dst.dynamicValues || []),
-			...diffArray(src.dynamicValues, oldComponentChild.dynamicValues),
-		];
-	}
-	if (src.innerHTML !== undefined && !deepEqual(src.innerHTML, oldComponentChild.innerHTML)) {
-		dst.innerHTML = src.innerHTML;
-	}
-	if (src.dataKey && !deepEqual(src.dataKey, oldComponentChild.dataKey)) {
-		dst.dataKey = src.dataKey;
-	}
-	if (src.visibilityCondition && !deepEqual(src.visibilityCondition, oldComponentChild.visibilityCondition)) {
-		dst.visibilityCondition = src.visibilityCondition;
-	}
-};
-
-const indexChildrenByRefId = (children: Block[] | undefined): Map<string, Block> => {
-	const map = new Map<string, Block>();
-	(children || []).forEach((child) => {
-		if (child.referenceBlockId) {
-			map.set(child.referenceBlockId, child);
-		}
-	});
-	return map;
-};
-
-const indexChildrenByBlockId = (children: Block[] | undefined): Map<string, Block> => {
-	const map = new Map<string, Block>();
-	(children || []).forEach((child) => {
-		if (child.blockId) {
-			map.set(child.blockId, child);
-		}
-	});
-	return map;
-};
-
-// Rebuild a component instance's children from the new version's structure, preserving user
-// overrides on matched children (matched by referenceBlockId via three-way diff against the old
-// version). Children with no counterpart in the new version are dropped — component subtrees
-// can't host detached blocks. Nested component instances fall back to resetWithComponent.
-function rebuildWithComponent(
-	block: Block,
-	componentId: string,
-	componentChildren: Block[],
-	oldChildrenByRefId: Map<string, Block>,
-	oldComponentChildrenByBlockId: Map<string, Block>,
-) {
-	block = toRaw(block);
-	block.children.splice(0, block.children.length);
-	componentChildren.forEach((componentChild) => {
-		const fresh = getBlockCopy(componentChild);
-		fresh.isChildOfComponent = componentId;
-		fresh.referenceBlockId = componentChild.blockId;
-		if (!fresh.extendedFromComponent) {
-			fresh.componentVersion = block.componentVersion;
-		}
-		const matched = oldChildrenByRefId.get(componentChild.blockId);
-		const oldComponentChild = oldComponentChildrenByBlockId.get(componentChild.blockId);
-		if (matched && oldComponentChild && !fresh.extendedFromComponent) {
-			copyUserOverrides(matched, fresh, oldComponentChild);
-		}
-		const childBlock = block.addChild(fresh, null, false);
-		if (componentChild.extendedFromComponent) {
-			const nestedComponent = childBlock.referenceComponent;
-			if (nestedComponent) {
-				resetWithComponent(childBlock, componentChild.extendedFromComponent, nestedComponent.children, false);
-			}
-		} else {
-			const nestedOld = indexChildrenByRefId(matched?.children);
-			const nestedOldComp = indexChildrenByBlockId(oldComponentChild?.children);
-			rebuildWithComponent(childBlock, componentId, componentChild.children, nestedOld, nestedOldComp);
-		}
-	});
-}
-
-function findComponentBlock(blockId: string, blocks: Block[]): Block | null {
-	for (const block of blocks) {
-		if (block.referenceBlockId === blockId) {
-			return block;
-		}
-		if (block.children) {
-			const found = findComponentBlock(blockId, block.children);
-			if (found) {
-				return found;
-			}
-		}
-	}
-	return null;
-}
-
-function resetBlock(
-	block: Block | BlockOptions,
-	resetChildren: boolean = true,
-	resetOverrides: boolean = true,
-) {
-	block.blockId = generateId();
-	if (resetOverrides) {
-		delete block.innerHTML;
-		delete block.element;
-		block.baseStyles = {};
-		block.rawStyles = {};
-		block.mobileStyles = {};
-		block.tabletStyles = {};
-		block.attributes = {};
-		block.customAttributes = {};
-		block.classes = [];
-		block.dataKey = null;
-		block.dynamicValues = [];
-		block.props = {};
-	}
-
-	if (resetChildren) {
-		block.children?.forEach((child) => {
-			resetBlock(child, resetChildren, !Boolean(child.extendedFromComponent));
-		});
-	}
-}
-
-export function findBlockInTree(blockId: string, blocks: Block[]): Block | null {
-	for (const block of blocks) {
-		if (block.blockId === blockId) {
-			return block;
-		}
-		if (block.children) {
-			const found = findBlockInTree(blockId, block.children);
-			if (found) {
-				return found;
-			}
-		}
-	}
-	return null;
 }
 
 export default Block;
